service/pyproject.toml
[project]
name = "service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi-structlog>=0.5.0",
    "fastapi>=0.115.2",
    "numpy>=2.1.2",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "python-multipart>=0.0.12",
    "structlog>=24.4.0",
    "torch>=2.4.1",
    "torchvision>=0.19.1",
    "uvicorn>=0.32.0",
]
